Minimum number of jumps
in this problem Given an array of N integers arr[] where each element represents the max number of steps that can be made forward from that element. Find the minimum number of jumps to reach the end of the array (starting from the first element). If an element is 0, then you cannot move through that element.

Expected Time Complexity: O(N)
Expected Space Complexity: O(1)

Hope you enjoy it..
thankyou,...