Find a triplet that sum to a given value
Find a triplet that sum to a given value:-
Given an array and a value, find if there is a triplet in array whose sum is equal to the given value. If there is such a triplet present in array, then print the triplet and return true. Else return false.

Hope this solution is help you..
Thankyou..