Median of two sorted arrays of different sizes
Median of two sorted arrays of different sizes:- 
Given two sorted arrays, a[] and b[], the task is to find the median of these sorted arrays, in O(log n + log m) time complexity, when n is the number of elements in the first array, and m is the number of elements in the second array.
This is an extension of median of two sorted arrays of equal size problem. Here we handle arrays of unequal size also.

It is a gfg problem solution hope it help you,.
Thankyou.