Split the binary string into substrings
Split the binary string into substrings with equal number of 0s and 1s:-
Given a binary string str of length N, the task is to find the maximum count of consecutive substrings str can be divided into such that all the substrings are balanced i.e. they have equal number of 0s and 1s. If it is not possible to split str satisfying the conditions then print -1.


it is a gfg question .. 
Hope this solution is help you..