Convert a sentence into its equivalent mobile
Convert a sentence into its equivalent mobile numeric keypad sequenceL-
Given a sentence in the form of a string, convert it into its equivalent mobile numeric keypad sequence. 
Examples : 
 

Input : GEEKSFORGEEKS
Output : 4333355777733366677743333557777
For obtaining a number, we need to press a
number corresponding to that character for 
number of times equal to position of the 
character. For example, for character C, 
we press number 2 three times and accordingly.

Input : HELLO WORLD
Output : 4433555555666096667775553

Hope this solution will help you. 
Thankyou..