Smallest distinct window
Smallest distinct window:-
Given a string 's'. The task is to find the smallest window length that contains all the characters of the given string at least one time.
For eg. A = “aabcbcdbca”, then the result would be 4 as of the smallest window will be “dbca”.

hope this solution will help you...
Thankyou...