Smallest window in a string containing all
Smallest window in a string containing all the characters of another string:-
Given two strings S and P. Find the smallest window in the string S consisting of all the characters(including duplicates) of the string P.  Return "-1" in case there is no such window present. In case there are multiple such windows of same length, return the one with the least starting index. 

Hope this solution is help you..

Thankyou..