Transform One String to Another using Min
Transform One String to Another using Minimum Number :-Given two strings A and B, the task is to convert A to B if possible. The only operation allowed is to put any character from A and insert it at front. Find if it’s possible to convert the string. If yes, then output minimum no. of operations required for transformation.

hope this solution will help you..
Thankyou..