Find first node of loop in a linked list
Find first node of loop in a linked list:-Write a function findFirstLoopNode() that checks whether a given Linked List contains a loop. If the loop is present then it returns point to the first node of the loop. Else it returns NULL.

Hope this solution will help you ..
Thankyou..