Remove duplicate element from sorted Linked List
Remove duplicate element from sorted Linked List:- Given a singly linked list consisting of N nodes. The task is to remove duplicates (nodes with duplicate values) from the given list (if exists).
Note: Try not to use extra space. Expected time complexity is O(N). The nodes are arranged in a sorted way.

Hope this solution is work for you...