Clone a linked list with next and random pointer
Clone a linked list with next and random pointer in O(1) space:-Given a linked list having two pointers in each node. The first one points to the next node of the list, however, the other pointer is random and can point to any node of the list. Write a program that clones the given list in O(1) space, i.e., without any extra space. 
Hope this solution will help you..