First non-repeating character in a stream
First non-repeating character in a stream :-Given an input stream of A of n characters consisting only of lower case alphabets. The task is to find the first non repeating character, each time a character is inserted to the stream. If there is no such character then append '#' to the answer.

Hope this solution will work for you..