Top View of Binary Tree
Top View of Binary Tree :-Given below is a binary tree. The task is to print the top view of binary tree. Top view of a binary tree is the set of nodes visible when the tree is viewed from the top. For the given below tree

Hope this solution will help you..