Transform to Sum Tree
Transform to Sum Tree :-Given a Binary Tree of size N , where each node can have positive or negative values. Convert this to a tree where each node contains the sum of the left and right sub trees of the original tree. The values of leaf nodes are changed to 0.

Hope this solution will help you..