Print all k-sum paths in a binary tree
Print all k-sum paths in a binary tree:-A binary tree and a number k are given. Print every path in the tree with sum of the nodes in the path as k. 
A path can start from any node and end at any node and must be downward only, i.e. they need not be root node and leaf node; and negative numbers can also be there in the tree.

Hope this solution will help you..