Minimum Cost to cut a board into squares
Minimum Cost to cut a board into squares:-A board of length m and width n is given, we need to break this board into m*n squares such that cost of breaking is minimum. cutting cost for each edge will be given for the board. In short, we need to choose such a sequence of cutting such that cost is minimized. 

Hope this solution will help you..