Minimum Cost of ropes
Minimum Cost of ropes :-There are given N ropes of different lengths, we need to connect these ropes into one rope. The cost to connect two ropes is equal to sum of their lengths. The task is to connect the ropes with minimum cost.

Hope this solution will help you..