Longest Possible Route in a Matrix with Hurdles
Longest Possible Route in a Matrix with Hurdles:-Given an M x N matrix, with a few hurdles arbitrarily placed, calculate the length of the longest possible route possible from source to a destination within the matrix. We are allowed to move to only adjacent cells which are not hurdles. The route cannot contain any diagonal moves and a location once visited in a particular path cannot be visited again.


Hope this solution will help you..