Design a stack with operations on middle element
Design a stack with operations on middle element:-How to implement a stack which will support following operations in O(1) time complexity? 
1) push() which adds an element to the top of stack. 
2) pop() which removes an element from top of stack. 
3) findMiddle() which will return middle element of the stack. 
4) deleteMiddle() which will delete the middle element. 
Push and pop are standard stack operations. 
The important question is, whether to use a linked list or array for the implementation of the stack? 
Please note that we need to find and delete the middle element. Deleting an element from the middle is not O(1) for array. Also, we may need to move the middle pointer up when we push an element and move down when we pop(). In a singly linked list, moving the middle pointer in both directions is not possible. 
The idea is to use Doubly Linked List (DLL). We can delete the middle element in O(1) time by maintaining mid pointer. We can move mid pointer in both directions using previous and next pointers. 
Following is implementation of push(), pop() and findMiddle() operations. Implementation of deleteMiddle() is left as an exercise. If there are even elements in stack, findMiddle() returns the second middle element. For example, if stack contains {1, 2, 3, 4}, then findMiddle() would return 3.


Hope this solution will help youl..