Knapsack with Duplicate Items
Knapsack with Duplicate Items :-Given a set of N items, each with a weight and a value, represented by the array w[] and val[] respectively. Also, a knapsack with weight limit W.
The task is to fill the knapsack in such a way that we can get the maximum profit. Return the maximum profit.


Hope this solution will help you..